requirements.txt
aniso8601==9.0.1
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
defusedxml==0.7.1
Flask==2.3.2
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-RESTful==0.3.9
fonttools==4.39.4
fpdf2==2.7.4
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
Pillow==9.5.0
pycparser==2.21
PyJWT==2.7.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.30.0
six==1.16.0
urllib3==2.0.2
Werkzeug==2.3.4
